Introduce the DXL Historian service
The DXL Historian service is a Daemon like service written in Python
that passively listens to OpenDXL event channels and records any
messages that it sees to a more persistent storage for analysis or 
other purposes.

In its current state the service is distributed as a Python package and
only has a single persistent recorder available (Elasticsearch). For
instructions on how to install, configure and operate the service,
refer to the `README.md` in the service directory. The service is 
expected to be managed using systemd. An accompanying unit file for the 
service can be found in `./infra/dxlhistorian.service`. Note that while
the service is capable of running as a non privileged user, the
installing party is responsible for enforcing that in the unit.

Utility scripts enabling simplistic installation and uninstallation of
the service can be found in the `./infra/scripts` folder. These scripts
still except a degree of "knowing what you are doing".