Introduce the DXL GRR service
The DXL GRR service is a rather thin wrapper around the GRR Rapid
Response server's API. The service itself is based on the HTTP
connector made available as part of the GRR project.

In general, the service itself is rather similar to the DXLHistorian
service with regard to the the non-functional requirements.
Functionality wise, it offers a number of topics corresponding to
endpoints available from the API. Note that the list of available
endpoints is compiled dynamically, by querying the configured GRR
server.

For further usage and installation information, refer to the README.md
file at the root of the project.