Introduce the DXL Wazuh service and update Robobluekit
The DXL Wazuh service offers a lightweight adapter to plug the Wazuh
API into a OpenDXL service fabric. The Service takes in JSON encoded
service calls and forwards them onto the Restful Wazuh API.

The list of available endpoints is based on programmatic parsing of the
generated Wazuh RESTful API documentation. This means that any changes
introduced to the API may be adapted to this adapter by merely updating
the JSON file in the repository.

Additionally, as part of this commit, a generic ServiceEndpointMonitor
was introduced into the Roboblue development kit as there's now more
than a single Service that provides monitoring information.